requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=4.8.0
python-dotenv>=0.19.0
cryptography>=3.4.7
psycopg2-binary>=2.9.1
django-cors-headers>=3.10.0
